verified token's scope

+++ django-tuttle/provider/management/commands/fetch_repositories.py
@@ -36,44 +36,51 @@ class Command(BaseCommand):
             # login on github account using user's token
             login = Github(token.token)
             user = login.get_user()
-            self.stdout.write('Saving repository info')
-            # delete user's deploy keys
-            DeployKey.objects.filter(repository__user=token.user).delete()
-            # get repository info of the logged user
-            for repo in user.get_repos():
-                params = {
-                    'name': repo.name,
-                    'owner': repo.owner.login,
-                    'organization': getattr(repo.organization, 'name', None),
-                    'is_private': repo.private,
-                    'is_user_admin': repo.permissions.admin,
-                    'user': token.user,
-                    'provider': provider,
-                }
-                try:
-                    Repository.objects.get_or_create(**params)
-                    self.stdout.write(repo.name)
 
-                except Repository.MultipleObjectsReturned:
-                    msg = 'More than 1 Repository with this params: %s ' % params
-                    logger.error(msg)
-                    raise CommandError(msg, 'Fix the database')
+            # check token scopes
+            scope_list = user.raw_headers['x-oauth-scopes']
+            if 'repo' in scope_list:
+                self.stdout.write('Saving repository info')
+                # delete user's deploy keys
+                DeployKey.objects.filter(repository__user=token.user).delete()
+                # get repository info of the logged user
+                for repo in user.get_repos():
+                    params = {
+                        'name': repo.name,
+                        'owner': repo.owner.login,
+                        'organization': getattr(repo.organization, 'name', None),
+                        'is_private': repo.private,
+                        'is_user_admin': repo.permissions.admin,
+                        'user': token.user,
+                        'provider': provider,
+                    }
+                    try:
+                        Repository.objects.get_or_create(**params)
+                        self.stdout.write(repo.name)
 
-                # user must be admin of his repository for get the deploy keys
-                if repo.permissions.admin:
-                    for key in repo.get_keys():
-                        params = {
-                            'title': key.title,
-                            'key': key.key,
-                            'repository': Repository.objects.get(name=repo.name),
-                        }
-                        try:
-                            DeployKey.objects.create(**params)
+                    except Repository.MultipleObjectsReturned:
+                        msg = 'More than 1 Repository with this params: %s ' % params
+                        logger.error(msg)
+                        raise CommandError(msg, 'Fix the database')
 
-                        except IntegrityError:
-                            msg = 'Error during the object creation'
-                            logger.error(msg)
-                            raise CommandError(msg)
+                    # user must be admin of his repository for get the deploy keys
+                    if repo.permissions.admin:
+                        for key in repo.get_keys():
+                            params = {
+                                'title': key.title,
+                                'key': key.key,
+                                'repository': Repository.objects.get(name=repo.name),
+                            }
+                            try:
+                                DeployKey.objects.create(**params)
+
+                            except IntegrityError:
+                                msg = 'Error during the object creation'
+                                logger.error(msg)
+                                raise CommandError(msg)
+
+            else:
+                raise CommandError('Your token has not REPO scope')
 
         except BadCredentialsException:
             logger.error('Login error on %s' % provider)
